Flytt typar for veileder og enhet til eiga fil slik at det er lettare å sjå at desse heng saman, og ikkje har noko med dei andre typane i model-interfaces å gjere
