+++ src/hooks/redux/use-innlogget-ident.tsx
@@ -1,7 +1,7 @@
 import {useSelector} from 'react-redux';
 import {AppState} from '../../reducer';
 import {OrNothing} from '../../utils/types/types';
-import {VeilederModell} from '../../model-interfaces';
+import {VeilederModell} from '../../typer/enhet-og-veiledere-modeller';
 
 const selectIdent = (state: AppState) => state.innloggetVeileder.data;
